# convert only necessary items