Not so beautiful code, but faster in my machine
14,5 vs 21,8