@jinyus 2023-10-13
Hey, this is slower on my machine(linux):

old:

```
Benchmark 1: ./target/release/rust
Processing time (w/o IO): 17.207528ms
Processing time (w/o IO): 16.176368ms
Processing time (w/o IO): 16.818853ms
Processing time (w/o IO): 16.937236ms
Processing time (w/o IO): 16.341477ms
Processing time (w/o IO): 17.013924ms
Processing time (w/o IO): 16.145778ms
Processing time (w/o IO): 16.99758ms
Processing time (w/o IO): 16.434089ms
Processing time (w/o IO): 16.137396ms
Processing time (w/o IO): 16.413346ms
Processing time (w/o IO): 16.087388ms
Processing time (w/o IO): 16.725961ms
  Time (mean ± σ):      36.5 ms ±   1.2 ms    [User: 26.0 ms, System: 10.3 ms]
  Range (min … max):    34.3 ms …  38.0 ms    10 runs
```

new:

```
Benchmark 1: ./target/release/rust
Processing time (w/o IO): 20.354093ms
Processing time (w/o IO): 20.419461ms
Processing time (w/o IO): 20.083611ms
Processing time (w/o IO): 20.268753ms
Processing time (w/o IO): 20.192141ms
Processing time (w/o IO): 20.300529ms
Processing time (w/o IO): 20.192909ms
Processing time (w/o IO): 20.386919ms
Processing time (w/o IO): 20.251503ms
Processing time (w/o IO): 20.226221ms
Processing time (w/o IO): 20.285234ms
Processing time (w/o IO): 20.310725ms
Processing time (w/o IO): 20.2485ms
  Time (mean ± σ):      36.8 ms ±   0.3 ms    [User: 25.4 ms, System: 11.3 ms]
  Range (min … max):    36.4 ms …  37.3 ms    10 runs

```