@jinyus 2023-10-13
also slower in the vm:

```
Rust:

        Benchmark 1: ./target/release/rust
        Processing time (w/o IO): 28.40785ms
        Processing time (w/o IO): 28.562453ms
        Processing time (w/o IO): 28.36835ms
        Processing time (w/o IO): 28.421651ms
        Processing time (w/o IO): 28.247648ms
        Processing time (w/o IO): 28.482252ms
        Processing time (w/o IO): 28.212547ms
        Processing time (w/o IO): 28.32885ms
        Processing time (w/o IO): 28.334049ms
        Processing time (w/o IO): 28.35545ms
        Processing time (w/o IO): 28.42425ms
        Processing time (w/o IO): 28.471352ms
        Processing time (w/o IO): 28.486651ms
          Time (mean ± σ):      46.1 ms ±   0.7 ms    [User: 40.1 ms, System: 5.9 ms]
          Range (min … max):    45.3 ms …  47.5 ms    10 runs
         
Rust:

        Benchmark 1: ./target/release/rust
        Processing time (w/o IO): 454.727112ms
        Processing time (w/o IO): 454.060201ms
        Processing time (w/o IO): 453.868698ms
          Time (mean ± σ):     525.7 ms ±   3.1 ms    [User: 496.1 ms, System: 29.5 ms]
          Range (min … max):   523.5 ms … 527.9 ms    2 runs
         
Rust:

        Benchmark 1: ./target/release/rust
        Processing time (w/o IO): 4.059799039s
        Processing time (w/o IO): 4.06099245s
        Processing time (w/o IO): 4.062162371s
          Time (mean ± σ):      4.308 s ±  0.004 s    [User: 4.218 s, System: 0.081 s]
          Range (min … max):    4.305 s …  4.311 s    2 runs
```