@artgmrv 2023-10-13
@jinyus interesting! What version of rustc is used in the benchmark? (my is 1.72)